+++ src/slither/explorerTree.ts
@@ -1,9 +1,9 @@
 import * as vscode from "vscode";
-import * as config from "./config";
-import { Logger } from "./logger";
+import * as config from "../config";
+import { Logger } from "../logger";
 import * as slither from "./slither";
 import * as slitherResults from "./slitherResults";
-import * as extension from "./extension";
+import * as extension from "../extension";
 
 // Generic tree node implementation.
 export class ExplorerNode extends vscode.TreeItem {
