+++ src/slither/slitherResults.ts
@@ -1,6 +1,6 @@
 import * as vscode from "vscode";
 import * as path from "path";
-import { Logger } from "./logger";
+import { Logger } from "../logger";
 
 export interface SlitherCommandOutput {
   success: boolean;
